chnages to keep only protos, add gha and compilation
