--- .env.example
@@ -1,7 +1,0 @@
-POSTGRES_USER=hermez
-POSTGRES_HOST=localhost
-POSTGRES_DB=polygon-hermez
-POSTGRES_PASSWORD=polygon
-POSTGRES_PORT=5432
-PROOF_TIME=5000
-INPUT_TIME=10000
