--- .github/workflows/docker-push.yml
@@ -1,20 +1,0 @@
-on:
-  push:
-    branches:
-    - main
-
-jobs:
-  build:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v2
-      - name: Login to DockerHub
-        uses: docker/login-action@v1
-        with:
-          username: ${{ secrets.DOCKERHUB_USERNAME }}
-          password: ${{ secrets.DOCKERHUB_TOKEN }}
-      - name: Build docker
-        run: npm run build:docker
-      - name: Push docker image
-        run: npm run push:docker
