+++ .github/workflows/proto.yml
@@ -1,0 +1,28 @@
+name: gRPC proto files
+
+on:
+  push:
+    branches:
+      - '*'
+  pull_request:
+
+jobs:
+   build:
+    name: gRPC proto files
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - name: Install buf
+        run: |
+         BIN="/usr/local/bin" && \
+         VERSION="1.0.0-rc12" && \
+         curl -sSL \
+         "https://github.com/bufbuild/buf/releases/download/v${VERSION}/buf-$(uname -s)-$(uname -m)" \
+         -o "${BIN}/buf" && \
+         chmod +x "${BIN}/buf"
+      - name: Run proto linter
+        run: buf lint proto --config ./buf.yaml
+      - name: Setup
+        run: npm run setup
+      - name: Comile proto
+        run: npm run test
