--- proto/buf.yaml
@@ -1,7 +1,0 @@
-version: v1
-breaking:
-  use:
-    - FILE
-lint:
-  use:
-    - DEFAULT
