+++ test/mt.test.js
@@ -1,0 +1,22 @@
+const protoLoader = require('@grpc/proto-loader');
+const grpc = require('@grpc/grpc-js');
+
+describe('Compile mt proto', () => {
+    const PROTO_PATH = `${__dirname}/../proto/mt/v1/mt.proto`;
+
+    it('compile', async () => {
+        const packageDefinition = protoLoader.loadSync(
+            PROTO_PATH,
+            {
+                keepCase: true,
+                longs: String,
+                enums: String,
+                defaults: true,
+                oneofs: true,
+            },
+        );
+
+        // eslint-disable-next-line no-unused-expressions
+        grpc.loadPackageDefinition(packageDefinition).zkprover;
+    });
+});
